core: регрессии (ST-каскад/IRecipeInput-мирор) + TE clean-швы кластер (гейт 103->~95)
Регрессии: ST TwilightTreasureReplacer.generate каст AbstractContainerMenu->Container (мой aInventory-каскад);
RecipeInputOreDict/ItemStack мироры implements matches(ItemStack) (я добавил в IRecipeInput). TE-швы:
blockEntityChanged(x,y,z)->getBlockPos() (TE04Covers), isInvalid->isRemoved (TE02Adjacent), getInventory().
getCurrentItem->getMainHandItem (DataSwitch), NBT Double->float каст (ConnectorRendered), getDescriptionPacket
->getUpdatePacket neo Packet<ClientGamePacketListener> (TE03TicksAndSync/TE02Sync F-network).

+++ src/main/java/gregapi/tileentity/connectors/TileEntityBase10ConnectorRendered.java
@@ -59,7 +59,7 @@ public abstract class TileEntityBase10ConnectorRendered extends TileEntityBase09
 	@Override
 	public void readFromNBT2(CompoundTag aNBT) {
 		super.readFromNBT2(aNBT);
-		if (aNBT.contains(NBT_DIAMETER)) mDiameter = Math.max(PX_P[2], Math.min(PX_N[0], (float)aNBT.getDouble(NBT_DIAMETER).orElse(0D)));
+		if (aNBT.contains(NBT_DIAMETER)) mDiameter = Math.max(PX_P[2], Math.min(PX_N[0], (float)(double)aNBT.getDouble(NBT_DIAMETER).orElse(0D)));
 		if (aNBT.contains(NBT_TRANSPARENT)) mTransparent = aNBT.getBoolean(NBT_TRANSPARENT).orElse(false);
 		if (aNBT.contains(NBT_CONTACTDAMAGE)) mContactDamage = aNBT.getBoolean(NBT_CONTACTDAMAGE).orElse(false);
 		if (aNBT.contains(NBT_FOAMDRIED)) mFoamDried = aNBT.getBoolean(NBT_FOAMDRIED).orElse(false);
